User Login, Signup, Registration Part Done using Components. Moved the Auth/AuthController Controller File inside Http/Controller and Removed auth.js (don't need that anymore
